+++ src/com/intellij/idea/plugin/hybris/common/HybrisConstants.kt
@@ -61,15 +61,15 @@ object HybrisConstants {
     const val IMPEX_CONFIG_COMPLETE_PREFIX = "$IMPEX_CONFIG_PREFIX-"
     const val IMPEX_CATALOG_VERSION_ONLINE = "Online"
     const val IMPEX_CATALOG_VERSION_STAGED = "Staged"
-    const val IMPEX_CONSOLE_TITLE = "[y] ImpEx Console"
-    const val IMPEX_MONITOR_CONSOLE_TITLE = "[y] Monitor Console"
     const val IMPEX = "ImpEx"
     const val IMPEX_FILE_EXTENSION = "impex"
 
-
-    const val GROOVY_CONSOLE_TITLE = "[y] Groovy Console"
-    const val FLEXIBLE_SEARCH_CONSOLE_TITLE = "[y] FS Console"
-    const val SOLR_SEARCH_CONSOLE_TITLE = "[y] Solr search"
+    const val CONSOLE_TITLE_IMPEX = "[y] ImpEx Console"
+    const val CONSOLE_TITLE_IMPEX_MONITOR = "[y] Monitor Console"
+    const val CONSOLE_TITLE_GROOVY = "[y] Groovy Console"
+    const val CONSOLE_TITLE_FLEXIBLE_SEARCH = "[y] FS Console"
+    const val CONSOLE_TITLE_SOLR_SEARCH = "[y] Solr search"
+    const val CONSOLE_TITLE_POLYGLOT_QUERY = "[y] PolyglotQuery"
 
     const val ROOT_TAG_BUSINESS_PROCESS_XML = "process"
     const val ROOT_TAG_ITEMS_XML = "items"