+++ src/com/intellij/idea/plugin/hybris/impex/actions/ImpExOpenQueryAction.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -49,7 +49,7 @@ class ImpExOpenQueryAction : AnAction() {
 
         CopyFileToHybrisConsoleUtils.copyQueryToConsole(
             project,
-            HybrisConstants.IMPEX_CONSOLE_TITLE,
+            HybrisConstants.CONSOLE_TITLE_IMPEX,
             query
         )
     }