+++ src/com/intellij/idea/plugin/hybris/tools/remote/console/impl/HybrisImpexConsole.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2023 EPAM Systems <hybrisideaplugin@epam.com>
+ * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -42,7 +42,7 @@ import java.io.Serial
 import javax.swing.Icon
 import javax.swing.JPanel
 
-class HybrisImpexConsole(project: Project) : HybrisConsole(project, HybrisConstants.IMPEX_CONSOLE_TITLE, ImpexLanguage.getInstance()) {
+class HybrisImpexConsole(project: Project) : HybrisConsole(project, HybrisConstants.CONSOLE_TITLE_IMPEX, ImpexLanguage.getInstance()) {
 
     object MyConsoleRootType : ConsoleRootType("hybris.impex.shell", null)